feat: revamp list service

+++ models/list.go
@@ -4,7 +4,7 @@ type List struct {
 	BaseModel
 	BoardID    uint64 `gorm:"foreign_key"`
 	Name       string `gorm:"type:varchar(50)"`
-	Position   int
+	Position   int64
 	IsArchived bool
 	Cards      []Card  `gorm:"foreignKey:ListID"`
 	Watches    []Watch `gorm:"foreignKey:ListID"`
